Add data persistence for the canister
Use the preupgrade() system method and  create a stable array where to push the HashMap data before the upgrade of the canister and use the postupgrade() system method to retrieve the data from it after the upgrade. Make the stable array and HashMap private to be accessible only within the canister.